Fixed input and tested it
