+++ SluaghEngine/Window/Input.cpp
@@ -1,6 +1,7 @@
 #include "Input.h"
 #include "gainput\GainputInputDeviceMouse.h"
 #include "gainput\GainputInputDeviceKeyboard.h"
+#include "gainput\gainputInputMap.h"
 namespace SE {
 	namespace Window {
 		Input::Input()
@@ -25,39 +26,53 @@ namespace SE {
 			inputMap = new gainput::InputMap(inputManager);
 		}
 
-		void Input::HandleMSG(const MSG & 	msg)
+		void Input::HandleMSG(MSG	msg)
 		{
 			return inputManager.HandleMessage(msg);
 		}
 
-		void Input::MapKeyToKeyboard(int actionKey, gainput::Key keyboardKey)
+		bool Input::MapKeyToKeyboard(int actionKey, Key keyboardKey)
 		{
-			inputMap->MapBool(actionKey, keyboardId, keyboardKey);
+			inputMap->MapBool(actionKey, keyboardId, (gainput::Key)keyboardKey);
+			return inputMap->IsMapped(actionKey);
 		}
 
-		void Input::MapKeyToMouse(int actionKey, gainput::MouseButton mouseButton)
+		bool Input::MapKeyToMouse(int actionKey, MouseButton mouseButton)
 		{
-			inputMap->MapBool(actionKey, mouseId, mouseButton);
+			inputMap->MapBool(actionKey, mouseId, (gainput::MouseButton)mouseButton);
+			return inputMap->IsMapped(actionKey);
 		}
 		
 		keyState Input::GetActionKeyState(int actionKey)
 		{
-			if (inputMap->GetBool(actionKey) == false)
+			if (inputMap->GetBoolPrevious(actionKey) == false)
 			{
-				return key_up;
-			}
-			else
-			{
-				if (inputMap->GetBoolIsNew(actionKey) == true)
+				if (inputMap->GetBoolIsNew(actionKey) == false)
+				{
+					return key_up;
+				}
+				else
 				{
 					return key_pressed;
 				}
-				else if (inputMap->GetBoolPrevious(actionKey) == true)
+			}
+			else
+			{
+				if (inputMap->GetBool(actionKey) == true)
 				{
 					return key_down;
 				}
+				else
+				{
+					return key_up;
+				}
 			}
 			return key_up;
+		}
+
+		void Input::Update()
+		{
+			inputManager.Update();
 		}
 	}	//namespace Window
 }	//namespace SE
