--- SluaghEngine/Window/KeyState.h
@@ -1,16 +1,0 @@
-#pragma once
-#ifndef SE_WINDOW_KEYSTATE_H
-#define SE_WINDOW_KEYSTATE_H
-
-namespace SE {
-	namespace Window {
-		enum keyState
-		{
-			key_up,
-			key_pressed,
-			key_down
-		};
-
-	}	//namespace Window
-}	//namespace SE
-#endif	//SE_WINDOW_KEYSTATE_H
