+++ SluaghEngine/Window/Window.cpp
@@ -4,6 +4,7 @@ namespace SE {
 	namespace Window {
 		bool Window::Initialise()
 		{		
+			windowMessage.message = WM_NULL;
 			input.InitInput();
 			bool initS = display.InitDisplay();
 			if (initS == false)
@@ -11,14 +12,14 @@ namespace SE {
 			return initS;
 		}
 
-		void Window::MapKeyToKeyboard(int actionKey, gainput::Key keyboardKey)
+		bool Window::MapKeyToKeyboard(int actionKey, Key keyboardKey)
 		{
-			input.MapKeyToKeyboard(actionKey, keyboardKey);
+			return input.MapKeyToKeyboard(actionKey, keyboardKey);
 		}
 
-		void Window::MapKeyToMouse(int actionKey, gainput::MouseButton mouseButton)
+		bool Window::MapKeyToMouse(int actionKey, MouseButton mouseButton)
 		{
-			input.MapKeyToMouse(actionKey, mouseButton);
+			return input.MapKeyToMouse(actionKey, mouseButton);
 		}
 
 		const void* Window::GethWnd()
@@ -26,9 +27,29 @@ namespace SE {
 			return display.GethWnd(); 
 		}
 
-		void Window::HandleMSG(const MSG & 	msg)
+		bool Window::HandleMSG()
 		{ 
-			input.HandleMSG(msg); 
+			while (windowMessage.message != WM_QUIT)
+			{
+
+				// Call every frame
+				input.Update();
+
+				if (PeekMessage(&windowMessage, NULL, NULL, NULL, PM_REMOVE)) {
+
+					TranslateMessage(&windowMessage);
+					DispatchMessage(&windowMessage);
+
+					input.HandleMSG(windowMessage);
+				}
+
+				// If there are no messages to handle, the application will continue by running a frame
+				else 
+				{
+					return true;
+				}
+			} 
+			return false;
 		}
 
 		keyState Window::GetActionKeyState(int actionKey)
