+++ SluaghEngine/Window/Window.vcxproj
@@ -19,11 +19,11 @@
     </ProjectConfiguration>
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="..\includes\Window\Display.h" />
+    <ClInclude Include="..\includes\Window\Input.h" />
     <ClInclude Include="..\includes\Window\InterfaceWindow.h" />
+    <ClInclude Include="..\includes\Window\KeyState.h" />
     <ClInclude Include="..\includes\Window\Window.h" />
-    <ClInclude Include="Display.h" />
-    <ClInclude Include="Input.h" />
-    <ClInclude Include="KeyState.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Display.cpp" />
