+++ SluaghEngine/Window/Window.vcxproj.filters
@@ -18,16 +18,16 @@
     <ClInclude Include="..\includes\Window\InterfaceWindow.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Display.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="Input.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="..\includes\Window\Window.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="KeyState.h">
+    <ClInclude Include="..\includes\Window\KeyState.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="..\includes\Window\Input.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="..\includes\Window\Display.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
