+++ SluaghEngine/includes/Window/Input.h
@@ -2,7 +2,8 @@
 #pragma once
 #ifndef SE_WINDOW_INPUT_H
 #define SE_WINDOW_INPUT_H
-
+#include "Windows.h"
+#include "Windowsx.h"
 
 #if defined(DEBUG)
 #pragma comment (lib, "GainputD.lib")
@@ -37,7 +38,7 @@ namespace SE {
 			* @param[in] msg Windows message 
 			*
 			*/
-			void HandleMSG(const MSG & 	msg);
+			void HandleMSG(MSG msg);
 			/**
 			* @brief	Maps keyboardKey to actionKey
 			*
@@ -45,8 +46,12 @@ namespace SE {
 			*
 			* @param[in] keyboardKey Value of the declared mouseButton
 			*
+			* @retval true Returns true if mapping suceded
+			*
+			* @retval false Returns false if mapping failed
+			*
 			*/
-			void MapKeyToKeyboard(int actionKey, gainput::Key keyboardKey);
+			bool MapKeyToKeyboard(int actionKey, Key keyboardKey);
 			/**
 			* @brief	Maps mouseButton to actionKey
 			*
@@ -54,8 +59,12 @@ namespace SE {
 			*
 			* @param[in] keyboardKey Value of the declared mouseButton
 			*
+			* @retval true Returns true if mapping suceded
+			*
+			* @retval false Returns false if mapping failed
+			*
 			*/
-			void MapKeyToMouse(int actionKey, gainput::MouseButton mouseButton);
+			bool MapKeyToMouse(int actionKey, MouseButton mouseButton);
 			/**
 			* @brief	Passes actionKey to get it's state
 			*
@@ -69,6 +78,11 @@ namespace SE {
 			*
 			*/
 			keyState GetActionKeyState(int actionKey);
+			/**
+			* @brief	Updates the inputManager
+			*
+			*/
+			void Update();
 		private:
 			int width = 640;
 			int height = 480;
